IT LIIIIVES. Server can compile and run now, connections can be made but chargen fails because mob bodys are not created and handled yet.
git-svn-id: svn://svn.code.sf.net/p/espress/code@8 62034c39-b87d-41b6-b390-722936910da7

+++ trunk/com/planet_ink/coffee_mud/Items/interfaces/Coins.java
@@ -22,9 +22,9 @@ public interface Coins extends Item
 	public long getNumberOfCoins();
 	public void setNumberOfCoins(long number);
 	public boolean putCoinsBack();
-	public double getDenomination();
-	public void setDenomination(double valuePerCoin);
-	public double getTotalValue();
+	public long getDenomination();
+	public void setDenomination(long valuePerCoin);
+	public long getTotalValue();
 	public String getCurrency();
 	public void setCurrency(String named);
 }
